Override isInstalled behavior
