+++ bin/schematic
@@ -41,10 +41,10 @@ require __DIR__.'/../../../autoload.php';
 // Disable the PHP include path
 Yii::$enableIncludePath = false;
 
-require_once __DIR__.'/../consolecommands/SchematicConsoleApp.php';
-
 // Because CHttpRequest is one of those stupid Yii files that has multiple classes defined in it.
 require_once CRAFT_APP_PATH.'framework/web/CHttpRequest.php';
+
+require_once __DIR__.'/../consolecommands/SchematicConsoleApp.php';
 
 Yii::setPathOfAlias('app', CRAFT_APP_PATH);
 Yii::setPathOfAlias('plugins', CRAFT_PLUGINS_PATH);
